Create MagicSquare_Algo
We define a magic square to be an n*n matrix of distinct positive integers from 1 to n*n where the sum of any row, column, or diagonal 
of length n is always equal to the same number: the magic constant.You will be given a 3*3 matrix  of integers in the inclusive range [1,9]. 
We can convert any digit a to any other digit b in the range [1,9] at cost of |a-b|. Given s, convert it into a magic square at minimal cost. 
Print this cost on a new line.